Replace contains() logic with a binary search
This changes the logic for contains() from using a linear search to
using a binary search through the elements in the list. In theory, this
should give faster expected time to find where an element is (or where
it should go on insert). The where-it-should-go functionality of the new
find_index_for() function will be useful in implementing functions
later, such as insert().